langchain/chains/extended/requirements.txt
langchain-groq==0.1.9
streamlit==1.37.0
python-dotenv>=1.0.1